requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
python-telegram-bot>=20.0
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic>=2.0.0
httpx>=0.27.0
